refactor(test): extract TestLogger from BaseHistoryRepositoryTest
