+++ src/test/java/com/youssef/GridPulse/domain/base/BaseHistoryRepositoryTest.java
@@ -9,7 +9,6 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.Duration;
 import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
@@ -274,31 +273,3 @@ public abstract class BaseHistoryRepositoryTest<H extends BaseHistoryEntity, ID 
     }
 }
 
-// Utility class for test logging
-class TestLogger {
-
-    public static void logSuiteStart(Class<?> testClass) {
-        System.out.println("\n⭐ " + testClass.getSimpleName() + " Test Execution Started");
-        System.out.println("⏰ Start Time: " + Instant.now());
-        System.out.println("-".repeat(60));
-    }
-
-    public static void logSuiteEnd(Class<?> testClass, Instant startTime) {
-        Instant endTime = Instant.now();
-        long duration = Duration.between(startTime, endTime).toMillis();
-
-        System.out.println("-".repeat(60));
-        System.out.println("🏁 " + testClass.getSimpleName() + " Test Execution Completed");
-        System.out.println("⏰ End Time: " + endTime);
-        System.out.println("⏱️  Total Duration: " + duration + "ms");
-        System.out.println("=".repeat(60));
-    }
-
-    public static void logTestStart(int testNumber) {
-        System.out.println("📋 Test " + testNumber + " - Setting up...");
-    }
-
-    public static void logTestEnd(int testNumber) {
-        System.out.println("✅ Test " + testNumber + " - Completed");
-    }
-}
